+++ doc/otsdaq_mu2e_tracker.org
@@ -171,6 +171,12 @@ And that no TIMEOUT should be observed on the very first event after a DTC_Reset
 * [[file:control_roc.org][control_ROC.py and its commands]]
 * [[file:./dtc_gui.org][DtcGui interface]]
 * LV/HV monitoring: [[file:lvhv_monitoring.org]]
+* programming DTCs (command line)
+#+begin_src
+/home/mu2ehwdev/program_both_DTCs_AL9.sh   /home/mu2ehwdev/DTC_Firmware/DTC2025Oct24_13_22.1_6ROC/DTC.bit
+/home/mu2ehwdev/program_one_FPGA_AL9.sh  0 /home/mu2ehwdev/DTC_Firmware/DTC2025Oct24_13_22.1_6ROC/DTC.bit # PCIE address
+/home/mu2ehwdev/program_one_FPGA_AL9.sh  N /home/mu2ehwdev/DTC_Firmware/DTC2025Oct24_13_22.1_6ROC/DTC.bit # JTAG index N
+#+end_src
 * slow control of DTCs                                                       
 ** DTC_0 FPGA:  4 parameters                                                 
   -- 0x9010 : temperature
