+++ otsdaq-mu2e-tracker/ArtModules/TrackerDQM_module.hh
@@ -30,8 +30,8 @@
 
 #include "artdaq-core/Data/Fragment.hh"
 
-// #include "artdaq-core-mu2e/Overlays/Decoders/TrackerDataDecoder.hh"
-#include "artdaq-core-mu2e/Data/TrackerDataDecoder.hh"
+#include "artdaq-core-mu2e/Overlays/Decoders/TrackerDataDecoder.hh"
+// #include "artdaq-core-mu2e/Data/TrackerDataDecoder.hh"
 #include "artdaq-core-mu2e/Overlays/DTCEventFragment.hh"
 #include "artdaq-core-mu2e/Overlays/FragmentType.hh"
 //-----------------------------------------------------------------------------
