+++ otsdaq-mu2e-tracker/Ui/DtcInterface.cc
@@ -12,8 +12,8 @@
 #include "iostream"
 #include "vector"
 
-// #include "artdaq-core-mu2e/Overlays/Decoders/TrackerDataDecoder.hh"
-#include "artdaq-core-mu2e/Data/TrackerDataDecoder.hh"
+#include "artdaq-core-mu2e/Overlays/Decoders/TrackerDataDecoder.hh"
+// #include "artdaq-core-mu2e/Data/TrackerDataDecoder.hh"
 
 #include "DtcInterface.hh"
 #include "TString.h"    // includes ROOT's Form
